+++ uni_modules/uni-file-picker/components/uni-file-picker/uni-file-picker.vue
@@ -278,7 +278,6 @@
 						files.push(Object.assign({}, v))
 					}
 				})
-			
 				this.uploadFiles(files)
 			},
 			async setValue(newVal, oldVal) {
@@ -294,7 +293,7 @@
 						v.fileID = url
 						v.url = await this.getTempFileURL(url)
 					}
-					v.path = v.url
+					if(v.url) v.path = v.url
 					return v
 				}
 				if (this.returnType === 'object') {
@@ -439,7 +438,7 @@
 				for (let i = 0; i < res.length; i++) {
 					const item = res[i]
 					const index = item.uuid ? this.files.findIndex(p => p.uuid === item.uuid) : item.index
-					
+
 					if (index === -1 || !this.files) break
 					if (item.errMsg === 'request:fail') {
 						this.files[index].url = item.path
@@ -457,7 +456,7 @@
 						}else{
 							this.files[index].url = item.url
 						}
-						
+
 						this.files[index].status = 'success'
 						this.files[index].progress += 1
 						successData.push(this.files[index])
@@ -521,7 +520,7 @@
 					this.setEmit()
 				})
 			},
-			
+
 			/**
 			 * 获取文件名和后缀
 			 * @param {Object} name
@@ -542,8 +541,7 @@
 				let data = []
 				if (this.returnType === 'object') {
 					data = this.backObject(this.files)[0]
-					this.localValue = {}
-					Object.assign(this.localValue, data)
+					this.localValue = data?data:null
 				} else {
 					data = this.backObject(this.files)
 					if (!this.localValue) {