+++ uni_modules/uni-list/components/uni-list-chat/uni-list-chat.vue
@@ -218,6 +218,10 @@
 </script>
 
 <style lang="scss" scoped>
+	$uni-font-size-lg:16px;
+	$uni-spacing-row-sm: 5px;
+	$uni-spacing-row-base: 10px;
+	$uni-spacing-row-lg: 15px;
 	$background-color: #fff;
 	$divide-line-color: #e5e5e5;
 	$avatar-width: 45px;