+++ uni_modules/uni-list/components/uni-list/uni-list.vue
@@ -55,6 +55,8 @@ export default {
 };
 </script>
 <style lang="scss" scoped>
+$uni-bg-color:#ffffff;
+$uni-border-color:#e5e5e5;
 .uni-list {
 	/* #ifndef APP-NVUE */
 	display: flex;