+++ uni_modules/uni-pagination/components/uni-pagination/uni-pagination.vue
@@ -257,6 +257,7 @@
 </script>
 
 <style lang="scss" scoped>
+	$uni-primary: #2979ff;
 	.uni-pagination {
 		/* #ifndef APP-NVUE */
 		display: flex;
@@ -281,7 +282,7 @@
 		/* #endif */
 		padding: 0 8px;
 		line-height: 30px;
-		font-size: $uni-font-size-base;
+		font-size: 12px;
 		position: relative;
 		background-color: #F0F0F0;
 		flex-direction: row;
@@ -298,13 +299,13 @@
 		/* #ifndef APP-NVUE */
 		display: flex;
 		/* #endif */
-		font-size: $uni-font-size-base;
+		font-size: 12px;
 		position: relative;
 		flex-direction: row;
 		justify-content: center;
 		align-items: center;
 		text-align: center;
-		color: #0F1214;
+		color: #666;
 		font-size: 12px;
 	}
 
@@ -318,8 +319,8 @@
 		align-items: center;
 		height: 30px;
 		line-height: 30px;
-		font-size: $uni-font-size-base;
-		color: $uni-text-color;
+		font-size: 12px;
+		color: #666;
 		margin: 0 5px;
 	}
 
@@ -333,7 +334,7 @@
 		text-align: center;
 		line-height: 30px;
 		// border: 1px red solid;
-		color: #666;
+		color: #999;
 		border-radius: 4px;
 		// border-width: 1px;
 		// border-style: solid;
@@ -365,16 +366,16 @@
 
 	.uni-pagination--hover {
 		color: rgba(0, 0, 0, 0.6);
-		background-color: $uni-bg-color-hover;
+		background-color: #eee;
 	}
 
 	.tag--active:hover {
-		color: $uni-color-primary;
+		color: $uni-primary;
 	}
 
 	.page--active {
 		color: #fff;
-		background-color: $uni-color-primary;
+		background-color: $uni-primary;
 	}
 
 	.page--active:hover {