+++ uni_modules/uni-transition/components/uni-transition/uni-transition.vue
@@ -112,7 +112,7 @@ export default {
 			if (obj.duration) {
 				this.durationTime = obj.duration
 			}
-			this.animation = createAnimation(Object.assign(this.config, obj))
+			this.animation = createAnimation(Object.assign(this.config, obj),this)
 		},
 		/**
 		 * 点击组件触发回调